Graphics/Engine
Cleanup/Refactor

+++ include/graphics/gs_mesh.h
@@ -16,7 +16,7 @@ typedef struct gs_mesh_t
 } gs_mesh_t;
 
 _force_inline
-gs_mesh_t gs_mesh_t_new( gs_vertex_attribute_type* layout_data, u32 layout_size, void* v_data, usize v_data_size )
+gs_mesh_t gs_mesh_t_new( gs_vertex_attribute_type* layout_data, usize layout_size, void* v_data, usize v_data_size )
 {
 	gs_graphics_i* gfx = gs_engine_instance()->ctx.graphics;
 	gs_mesh_t mesh = {0};
